running_mate_ai_agent_service/requirements.txt
fastapi
uvicorn
python-dotenv
aiokafka
requests